+++ src/utils/Index.js
@@ -2,3 +2,5 @@ export { default as checkKc } from './checkKc.js';
 export { default as grumblerCheck } from './grumblerCheck.js';
 export { default as sanitizedTime } from './sanitizedTime.js';
 export { default as totalLevelCheck } from './totalLevelCheck.js';
+export { default as collectionLogCheck } from './collectionLogCheck.js';
+export { default as petCheck } from './petCheck.js';
