+++ src/utils/grumblerCheck.js
@@ -1,4 +1,5 @@
 import * as Constants from '../constants.js';
+
 /**
  * Always check for the grumbler
  * @param {*} name
